Fully Fledged responsive testengine
This is a online test portal where students and teachers can login/register and take test/create test and get their result live. This is implemented using vanilla JS at frontend and Firebase online database as backend. This project can be scalable and completed with MVC framework.

This project can also be used in University / School by professor to check the performance of students.
Also it can be used by Company to check performance of employees.